chore/bug: fix menuID selector failure to render for first time on Home page
Due to the async nature of setting state, the businessUID set at App.js level was not available when the Home page first rendered.
Two things, i) add businessUID to useEffect in MenuIDSelector to cause it to rerender when the businessUID becomes available, and ii) display placeholder while wait for
businessUID to become available.
Finally, added menuID as prop to WebRTCSTT component to allow it create the Agent when start is pressed.
This finalizes https://www.pivotaltracker.com/story/show/186662397