Feature: Allow other components in the useAvatar context to set the avatar status
WebRTC component is responsible for handling speech to text. When speech
is first recieved the status needs to change to Listenign so that the
avatar can respond and give visual feedback to the user. Moving
avatarStatus to the useAvatar context and passing it down to the children
of the context allows this. Passing by ref rather than state change
avoids rerenders of the avatar.