Feature: Add conditional animation transitions
Let the animation at the time of the avatar state change determine
whether an animation in a composite animation should be scheudled. For
example, if the idle_config schedules the avatar to standup from
sitting, then this shouldnt be scheduled if the avatar is not sitting at
the time of the state change.